refactor(portfolio): main structure
